+++ CHANGELOG.md
@@ -10,10 +10,12 @@ Unreleased in the current development version (target v0.20.0):
 ClimateDT workflow modifications:
 
 AQUA core complete list:
+- eORCA025 nested grid for o25.1 and o26.1 support (#2459)
 - Safe dumping yaml with a temporary file and SoftFileLock (#2445)
 - Add filelock when modifying catalog sources (#2432)
 
 AQUA diagnostics complete list:
+- Ocean3D: removed old diagnostic files (#2467)
 - Teleconnections: ENSO better vmin/vmax for regression plots (#2453)
 - Ocean Stratification and Ocean Trend: full support for regions across the dateline (#2433)
 - Area selection full support for regions across the dateline (#2430, #2434)
@@ -28,6 +30,7 @@ ClimateDT workflow modifications:
 - New position for push_analysis `config.grouping.yaml` in `$AQUA/config/analysis`
 
 AQUA core complete list:
+- Centralise `save_figure` in OutputSaver class (#2425)
 - Allow longer time default for connection to dask cluster (#2420)
 - Safe parallel creation of area and weight files (#2412)
 - New grid lat-lon-r100 with explicit grid file, new default in DROP catgen (#2410, #2413)